backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
tensorflow==2.20.0
pillow==10.0.1
numpy==1.24.3